requirements.txt
lxml
numpy
newspaper3k
trafilatura
beautifulsoup4
requests
pandas
scikit-learn
scipy